requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
spacy>=3.7.0
pyspellchecker>=0.8.1
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=2.7.0
wordcloud>=1.9.3
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
openpyxl>=3.1.0
joblib>=1.3.0
